final version of the screenshot component
