+++ src/components/screenshot/index.js
@@ -1,0 +1,1 @@
+export * from './screenshot';
